requirements.txt
azure-core==1.9.0
azure-storage-blob==12.6.0
azure-storage-file-datalake==12.2.0
databricks-connect==7.3.4b0
matplotlib==3.3.3
numpy==1.19.3
pandas==1.1.4
jupyter==1.0.0